refactoring docker-compose (#1228)
