+++ .github/workflows/release.yml
@@ -13,14 +13,18 @@ jobs:
         uses: actions/checkout@v2
         with:
           fetch-depth: 0
+
       - name: Set up Go
         uses: actions/setup-go@v2
         with:
           go-version: 1.17
+
       - name: Get packr
         run: go get -u github.com/gobuffalo/packr/v2/packr2
+
       - name: Prepare
         run: git reset --hard
+
       - name: Run GoReleaser
         uses: goreleaser/goreleaser-action@v2
         with:
