+++ .github/workflows/test-e2e.yml
@@ -23,19 +23,30 @@ jobs:
     steps:
     - name: Checkout code
       uses: actions/checkout@v2
+
     - name: Install Go
       uses: actions/setup-go@v1
       with:
         go-version: ${{ matrix.go-version }}
       env:
         GOARCH: ${{ matrix.goarch }}
+
     - name: Login to DockerHub
       uses: docker/login-action@v1
       with:
         username: ${{ secrets.DOCKERHUB_USERNAME }}
         password: ${{ secrets.DOCKERHUB_TOKEN }}
+
+    - name: Build Docker
+      run: make build-docker
+
+    - name: Compile SCs
+      run: make compile-scs
+      working-directory: test
+
     - name: Test
       run: make test-e2e-group-${{ matrix.e2e-group }}
+      working-directory: test
 
   from-fork-test-e2e:
     if:
@@ -53,19 +64,30 @@ jobs:
       uses: actions/checkout@v2
       with:
         ref: 'refs/pull/${{ github.event.client_payload.pull_request.number }}/merge'
+
     - name: Install Go
       uses: actions/setup-go@v1
       with:
         go-version: ${{ matrix.go-version }}
       env:
         GOARCH: ${{ matrix.goarch }}
+
     - name: Login to DockerHub
       uses: docker/login-action@v1
       with:
         username: ${{ secrets.DOCKERHUB_USERNAME }}
         password: ${{ secrets.DOCKERHUB_TOKEN }}
+
+    - name: Build Docker
+      run: make build-docker
+
+    - name: Compile SCs
+      run: make compile-scs
+      working-directory: test
+
     - name: Test
-      run: CONFIG_MODE="test" make test-e2e-group-${{ matrix.e2e-group }}
+      run: make test-e2e-group-${{ matrix.e2e-group }}
+      working-directory: test
 
     # Update check run
     - uses: actions/github-script@v5
