+++ .github/workflows/test-from-prover.yml
@@ -40,5 +40,13 @@ jobs:
         username: ${{ secrets.DOCKERHUB_USERNAME }}
         password: ${{ secrets.DOCKERHUB_TOKEN }}
 
+    - name: Build Docker
+      run: make build-docker
+
+    - name: Compile SCs
+      run: make compile-scs
+      working-directory: test
+
     - name: Test
       run: make test-full-non-e2e
+      working-directory: test
