+++ .github/workflows/test-full-non-e2e.yml
@@ -22,21 +22,29 @@ jobs:
     steps:
     - name: Checkout code
       uses: actions/checkout@v2
+
     - name: Install Go
       uses: actions/setup-go@v1
       with:
         go-version: ${{ matrix.go-version }}
       env:
         GOARCH: ${{ matrix.goarch }}
+
     - name: Login to DockerHub
       uses: docker/login-action@v1
       with:
         username: ${{ secrets.DOCKERHUB_USERNAME }}
         password: ${{ secrets.DOCKERHUB_TOKEN }}
+
+    - name: Compile SCs
+      run: make compile-scs
+      working-directory: test
+
     - name: Test
       env:
         ZKPROVER_URI: 127.0.0.1
       run: make test-full-non-e2e
+      working-directory: test
 
   from-fork-test-full-non-e2e:
     if:
@@ -53,21 +61,32 @@ jobs:
       uses: actions/checkout@v2
       with:
         ref: 'refs/pull/${{ github.event.client_payload.pull_request.number }}/merge'
+
     - name: Install Go
       uses: actions/setup-go@v1
       with:
         go-version: ${{ matrix.go-version }}
       env:
         GOARCH: ${{ matrix.goarch }}
+
     - name: Login to DockerHub
       uses: docker/login-action@v1
       with:
         username: ${{ secrets.DOCKERHUB_USERNAME }}
         password: ${{ secrets.DOCKERHUB_TOKEN }}
+
+    - name: Build Docker
+      run: make build-docker
+
+    - name: Compile SCs
+      run: make compile-scs
+      working-directory: test
+
     - name: Test
       env:
         ZKPROVER_URI: 127.0.0.1
       run: make test-full-non-e2e
+      working-directory: test
 
     # Update check run
     - uses: actions/github-script@v5
