+++ Dockerfile
@@ -1,8 +1,6 @@
 # CONTAINER FOR BUILDING BINARY
 FROM golang:1.17 AS build
 
-ENV CGO_ENABLED=0
-
 # INSTALL DEPENDENCIES
 RUN go install github.com/gobuffalo/packr/v2/packr2@v2.8.3
 COPY go.mod go.sum /src/
@@ -16,6 +14,6 @@ RUN cd /src && make build
 # CONTAINER FOR RUNNING BINARY
 FROM alpine:3.16.0
 COPY --from=build /src/dist/zkevm-node /app/zkevm-node
-COPY --from=build /src/config/config.local.toml /app/config.local.toml
+COPY --from=build /src/config/environments/local/local.node.config.toml /app/example.config.toml
 EXPOSE 8123
 CMD ["/bin/sh", "-c", "/app/zkevm-node run"]
