+++ config/default.go
@@ -29,7 +29,7 @@ MaxConns = 200
 [Etherman]
 URL = "http://localhost:8545"
 L1ChainID = 1337
-PrivateKeyPath = "./test/test.keystore"
+PrivateKeyPath = "/pk/keystore"
 PrivateKeyPassword = "testonly"
 PoEAddr = "0x2279B7A0a67DB372996a5FaB50D91eAA73d2eBe6"
 MaticAddr = "0x5FbDB2315678afecb367f032d93F642f64180aa3"
