+++ config/environments/local/debug.node.config.toml
@@ -8,24 +8,24 @@ Outputs = ["stdout"]
 User = "state_user"
 Password = "state_password"
 Name = "state_db"
-Host = "zkevm-state-db"
+Host = "localhost"
 Port = "5432"
 EnableLog = false
-MaxConns = 200
+MaxConns = 10
 
 [PoolDB]
 User = "pool_user"
 Password = "pool_password"
 Name = "pool_db"
-Host = "zkevm-pool-db"
-Port = "5432"
+Host = "localhost"
+Port = "5433"
 EnableLog = false
-MaxConns = 200
+MaxConns = 10
 
 [Etherman]
-URL = "http://zkevm-mock-l1-network:8545"
+URL = "http://localhost:8545"
 L1ChainID = 1337
-PrivateKeyPath = "./test/test.keystore"
+PrivateKeyPath = "../config/example.keystore"
 PrivateKeyPassword = "testonly"
 PoEAddr = "0x2279B7A0a67DB372996a5FaB50D91eAA73d2eBe6"
 MaticAddr = "0x5FbDB2315678afecb367f032d93F642f64180aa3"
@@ -43,7 +43,7 @@ PercentageToIncreaseGasLimit = 10
 [RPC]
 Host = "0.0.0.0"
 Port = 8123
-MaxRequestsPerIPAndSecond = 5000
+MaxRequestsPerIPAndSecond = 10000
 SequencerNodeURI = ""
 BroadcastURI = "127.0.0.1:61090"
 DefaultSenderAddress = "0x1111111111111111111111111111111111111111"
@@ -51,13 +51,13 @@ DefaultSenderAddress = "0x1111111111111111111111111111111111111111"
 		User = "rpc_user"
 		Password = "rpc_password"
 		Name = "rpc_db"
-		Host = "zkevm-rpc-db"
-		Port = "5432"
+		Host = "localhost"
+		Port = "5434"
 		EnableLog = false
 		MaxConns = 10
 
 [Synchronizer]
-SyncInterval = "1s"
+SyncInterval = "5s"
 SyncChunkSize = 100
 TrustedSequencerURI = ""
 GenBlockNumber = 1
@@ -94,7 +94,7 @@ Type = "default"
 DefaultGasPriceWei = 1000000000
 
 [Provers]
-ProverURIs = ["zkevm-prover:50052"]
+ProverURIs = ["localhost:50052"]
 
 [MTServer]
 Host = "0.0.0.0"
@@ -102,10 +102,10 @@ Port = 50060
 StoreBackend = "PostgreSQL"
 
 [MTClient]
-URI = "zkevm-prover:50061"
+URI = "127.0.0.1:50061"
 
 [Executor]
-URI = "zkevm-prover:50071"
+URI = "127.0.0.1:50071"
 
 [BroadcastServer]
 Host = "0.0.0.0"
