+++ config/environments/public/public.node.config.toml
@@ -1,0 +1,60 @@
+[Log]
+Level = "debug"
+Outputs = ["stdout"]
+
+[StateDB]
+User = "state_user"
+Password = "state_password"
+Name = "state_db"
+Host = "zkevm-state-db"
+Port = "5432"
+EnableLog = false
+MaxConns = 200
+
+[PoolDB]
+User = "pool_user"
+Password = "pool_password"
+Name = "pool_db"
+Host = "zkevm-pool-db"
+Port = "5432"
+EnableLog = false
+MaxConns = 200
+
+[Etherman]
+URL = "https://geth_goerli.hermez.io"
+PrivateKeyPath = "/pk/keystore"
+PrivateKeyPassword = "testonly"
+PoEAddr = "0x14cB06e8dE2222912138F9a062E5a4d9F4821409"
+MaticAddr = "0x4701Aa9471d7bfAc765D87dcb1Ea6BB23AD32733"
+GlobalExitRootManagerAddr = "0x762d09Ed110ace4EaC94acbb67b1C35D16C3297b"
+
+[RPC]
+Host = "0.0.0.0"
+Port = 8545
+MaxRequestsPerIPAndSecond = 5000
+SequencerNodeURI = "https://public.zkevm-test.net:2083"
+BroadcastURI = "public-grpc.zkevm-test.net:61090"
+DefaultSenderAddress = "0x1111111111111111111111111111111111111111"
+	[RPC.DB]
+		User = "rpc_user"
+		Password = "rpc_password"
+		Name = "rpc_db"
+		Host = "zkevm-rpc-db"
+		Port = "5432"
+		EnableLog = false
+		MaxConns = 100
+
+[Synchronizer]
+SyncInterval = "2s"
+SyncChunkSize = 10000
+TrustedSequencerURI = ""
+
+[GasPriceEstimator]
+Type = "default"
+DefaultGasPriceWei = 1000000000
+
+[MTClient]
+URI = "zkevm-prover:50061"
+
+[Executor]
+URI = "zkevm-prover:50071"
