+++ config/environments/public/public.prover.config.json
@@ -1,6 +1,6 @@
 {
     "runProverServer": false,
-    "runProverServerMock": true,
+    "runProverServerMock": false,
     "runProverClient": false,
 
     "runExecutorServer": true,
