+++ db/scripts/init_prover_db.sql
