+++ test/scripts/cmd/compilesc/manager.go
