+++ test/scripts/cmd/dependencies.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/0xPolygonHermez/zkevm-node/scripts/cmd/dependencies"
+	"github.com/0xPolygonHermez/zkevm-node/test/scripts/cmd/dependencies"
 	"github.com/urfave/cli/v2"
 )
