+++ test/scripts/cmd/dependencies/files.go
