+++ test/scripts/cmd/dependencies/files_test.go
