+++ test/scripts/cmd/dependencies/testvectors.go
