+++ test/scripts/cmd/main.go
