+++ test/scripts/deploy_sc/main.go
