+++ test/scripts/init_network/main.go
